backend/requirements.txt
# Core Django packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Environment management
python-decouple==3.8

# API documentation
drf-spectacular==0.26.5

# HTTP requests for external APIs
requests==2.31.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0

# Additional utilities
django-filter==23.3
Pillow==10.0.1
python-dateutil==2.8.2

# Development dependencies
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.3.2

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
